Added a "layouts" module
