--- samples/basic/host/src/OfficevibeTab.tsx
@@ -1,10 +1,0 @@
-export function OfficevibeTab() {
-    return (
-        <>
-            <h2>Officevibe</h2>
-            <p style={{ backgroundColor: "purple", color: "white", width: "fit-content" }}>This tab is served by <code>@basic/host</code></p>
-        </>
-    );
-}
-
-export const Component = OfficevibeTab;
