+++ samples/basic/host/src/register.tsx
@@ -5,22 +5,6 @@ function registerRoutes(runtime: Runtime) {
         index: true,
         lazy: () => import("./HomePage.tsx")
     });
-
-    // Register federated tab.
-
-    runtime.registerRoute({
-        path: "/federated-tabs/officevibe",
-        lazy: () => import("./OfficevibeTab.tsx")
-    }, {
-        parentPath: "/federated-tabs"
-    });
-
-    runtime.registerNavigationItem({
-        $label: "Officevibe",
-        to: "/federated-tabs/officevibe"
-    }, {
-        menuId: "/federated-tabs"
-    });
 }
 
 export const registerHost: ModuleRegisterFunction<Runtime> = runtime => {
