+++ samples/basic/remote-module/src/dev/register.tsx
@@ -6,14 +6,4 @@ export const registerDev: ModuleRegisterFunction<Runtime> = runtime => {
         index: true,
         element: <DevHome />
     });
-
-    runtime.registerRoute({
-        path: "/federated-tabs",
-        lazy: () => import("@basic/shared/FederatedTabsLayout.tsx")
-    });
-
-    runtime.registerNavigationItem({
-        $label: "Tabs",
-        to: "/federated-tabs"
-    });
 };
