+++ samples/basic/shared/src/index.ts
@@ -2,5 +2,6 @@ export * from "./BackgroundColorContext.ts";
 export * from "./appContext.ts";
 export * from "./eventBus.ts";
 export * from "./isNetlify.ts";
+export * from "./layouts/registerLayouts.tsx";
 export * from "./session.ts";
