+++ samples/basic/shell/tsconfig.json
@@ -7,8 +7,7 @@
             "@squide/webpack-module-federation": ["../../../packages/webpack-module-federation/src/index.ts"],
             "@squide/webpack-module-federation/defineConfig.js": ["../../../packages/webpack-module-federation/src/defineConfig.js"],
             "@squide/fakes": ["../../../packages/fakes/src/index.ts"],
-            "@basic/shared": ["../shared/src/index.ts"],
-            "@basic/shared/FederatedTabsLayout.tsx": ["../shared/src/FederatedTabsLayout.tsx"]
+            "@basic/shared": ["../shared/src/index.ts"]
         }
     },
     "exclude": ["dist", "node_modules"]
