+++ samples/endpoints/remote-module/tsconfig.json
@@ -9,8 +9,7 @@
             "@squide/msw": ["../../../packages/msw/src/index.ts"],
             "@squide/fakes": ["../../../packages/fakes/src/index.ts"],
             "@endpoints/shell": ["../shell/src/index.ts"],
-            "@endpoints/shared": ["../shared/src/index.ts"],
-            "@endpoints/shared/FederatedTabsLayout.tsx": ["../shared/src/FederatedTabsLayout.tsx"]
+            "@endpoints/shared": ["../shared/src/index.ts"]
         }
     },
     "exclude": ["dist", "node_modules"]
