+++ samples/endpoints/shared/package.json
@@ -11,11 +11,6 @@
             "import": "./dist/index.js",
             "types": "./dist/index.d.ts",
             "default": "./dist/index.js"
-        },
-        "./FederatedTabsLayout.tsx": {
-            "import": "./dist/FederatedTabsLayout.js",
-            "types": "./dist/FederatedTabsLayout.d.ts",
-            "default": "./dist/FederatedTabsLayout.js"
         }
     },
     "scripts": {
@@ -26,7 +21,6 @@
     },
     "peerDependencies": {
         "@squide/react-router": "*",
-        "@tanstack/react-query": "*",
         "react": "*",
         "react-dom": "*",
         "react-error-boundary": "*",
