+++ samples/endpoints/shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./featureFlags.ts";
 export * from "./isNetlify.ts";
+export * from "./layouts/registerLayouts.tsx";
 export * from "./session.ts";
 export * from "./subscription.ts";
 export * from "./telemetryService.ts";
